quant_platform/requirements.txt
pandas
polars
pyarrow
numpy
scikit-learn
yfinance
requests
beautifulsoup4
ta
fredapi
praw
xgboost
optuna
torch
transformers
peft
trl
bitsandbytes
datasets
wandb
fastapi
uvicorn
pydantic
openai
schedule
ibapi
python-telegram-bot
matplotlib
seaborn
plotly
joblib
dask
pandas-market-calendars
duckdb
pyyaml
modal
